services/web/requirements.txt
Werkzeug==1.0.1
Flask_Cors==3.0.8
Flask==1.1.2
Flask_Mail==0.9.1
praw==7.1.0
Flask_SQLAlchemy==2.4.4
python-dotenv==0.14.0
gunicorn==20.0.4
